добавляет реализацию для быстрого перехода к случайному городу из компонента логинскрин"
